Cleanup and refractoring
EGL / OpenGL ES also now working
