+++ include/GLFW/glfw3.h
@@ -5208,5 +5208,3 @@ GLFWAPI VkResult glfwCreateWindowSurface(VkInstance instance, GLFWwindow* window
 #endif
 
 #endif /* _glfw3_h_ */
-
-extern int main();
