+++ src/android_window.c
@@ -29,46 +29,6 @@
 #include <android_native_app_glue.h>
 #include <android/log.h>
 
-struct android_app* app;
-
-void handle_cmd(struct android_app* _app, int32_t cmd) {
-    switch (cmd) {
-        case APP_CMD_INIT_WINDOW:
-            // The window is being shown, get it ready.
-            app = _app;
-            __android_log_print(ANDROID_LOG_INFO, "GLFW",
-                                "Window initialized");
-        default:
-            __android_log_print(ANDROID_LOG_INFO, "GLFW",
-                                "event not handled: %d", cmd);
-    }
-}
-
-// Android Entry Point
-void android_main(struct android_app *app) {
-    app->onAppCmd = handle_cmd;
-    pthread_t t;pthread_create(&t, NULL, &main, NULL); // Call the main entry point
-
-    while (1) {
-        int ident;
-        int events;
-        struct android_poll_source* source;
-
-        while ((ident=ALooper_pollAll(0, NULL, &events,(void**)&source)) >= 0) {
-
-            // Process this event.
-            if (source != NULL) {
-                source->process(app, source);
-            }
-
-            // Check if we are exiting.
-            if (app->destroyRequested != 0) {
-                return;
-            }
-        }
-    }
-
-}
 //////////////////////////////////////////////////////////////////////////
 //////                       GLFW platform API                      //////
 //////////////////////////////////////////////////////////////////////////
@@ -78,7 +38,6 @@ int _glfwPlatformCreateWindow(_GLFWwindow* window,
                               const _GLFWctxconfig* ctxconfig,
                               const _GLFWfbconfig* fbconfig)
 {
-    while (app == NULL); // Wait for the app to be initialized or the app will crash occasionally
     window->android.app = app;
 
     if (ctxconfig->client != GLFW_NO_API)
